+++ engine/common/components-tests/src/test/scala/pl/touk/nussknacker/engine/common/components/DecisionTableSpec.scala
@@ -6,6 +6,8 @@ import org.scalatest.freespec.AnyFreeSpec
 import org.scalatest.matchers.should.Matchers
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError
 import pl.touk.nussknacker.engine.api.context.ProcessCompilationError.ExpressionParserCompilationError
+import pl.touk.nussknacker.engine.api.generics.ExpressionParseError.TabularDataDefinitionParserErrorDetails
+import pl.touk.nussknacker.engine.api.generics.ExpressionParseError.TabularDataDefinitionParserErrorDetails.CellError
 import pl.touk.nussknacker.engine.build.{GraphBuilder, ScenarioBuilder}
 import pl.touk.nussknacker.engine.canonicalgraph.CanonicalProcess
 import pl.touk.nussknacker.engine.flink.test.FlinkSpec
@@ -96,7 +98,8 @@ trait DecisionTableSpec extends AnyFreeSpec with Matchers with ValidatedValuesDe
                 message = "There is no property 'years' in type: Record{DoB: LocalDate, age: Integer, name: String}",
                 nodeId = "decision-table",
                 fieldName = Some("Expression"),
-                originalExpr = "#ROW['years'] > #input.minAge"
+                originalExpr = "#ROW['years'] > #input.minAge",
+                details = None
               )
             )
           )
@@ -119,7 +122,8 @@ trait DecisionTableSpec extends AnyFreeSpec with Matchers with ValidatedValuesDe
                 message = "Wrong part types",
                 nodeId = "decision-table",
                 fieldName = Some("Expression"),
-                originalExpr = "#ROW['name'] > #input.minAge"
+                originalExpr = "#ROW['name'] > #input.minAge",
+                details = None
               )
             )
           )
@@ -142,10 +146,31 @@ trait DecisionTableSpec extends AnyFreeSpec with Matchers with ValidatedValuesDe
           errors should be(
             NonEmptyList.one(
               ExpressionParserCompilationError(
-                message = "Column has a 'java.lang.Object' type but the value 'John' cannot be converted to it.",
+                message = "Typing error in some cells",
                 nodeId = "decision-table",
                 fieldName = Some("Basic Decision Table"),
-                originalExpr = invalidColumnTypeDecisionTableJson.expression
+                originalExpr = invalidColumnTypeDecisionTableJson.expression,
+                details = Some(
+                  TabularDataDefinitionParserErrorDetails(
+                    List(
+                      CellError(
+                        columnName = "name",
+                        rowIndex = 0,
+                        errorMessage = "Column has 'Object' type but its value cannot be converted to the type."
+                      ),
+                      CellError(
+                        columnName = "name",
+                        rowIndex = 1,
+                        errorMessage = "Column has 'Object' type but its value cannot be converted to the type."
+                      ),
+                      CellError(
+                        columnName = "name",
+                        rowIndex = 2,
+                        errorMessage = "Column has 'Object' type but its value cannot be converted to the type."
+                      )
+                    )
+                  )
+                )
               )
             )
           )