+++ interpreter/src/main/scala/pl/touk/nussknacker/engine/expression/NullExpression.scala
@@ -1,11 +1,12 @@
 package pl.touk.nussknacker.engine.expression
 
 import pl.touk.nussknacker.engine.api.Context
+import pl.touk.nussknacker.engine.graph.expression.Expression.Language
 import pl.touk.nussknacker.engine.api.expression.{Expression => CompiledExpression}
 import pl.touk.nussknacker.engine.spel.SpelExpressionParser.Flavour
 
 case class NullExpression(original: String, flavour: Flavour) extends CompiledExpression {
-  override def language: String = flavour.languageId
 
+  override val language: Language                                      = flavour.languageId
   override def evaluate[T](ctx: Context, globals: Map[String, Any]): T = null.asInstanceOf[T]
 }